LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/contain-intrinsic-size/auto-008-expected.txt


FAIL requestAnimationFrame assert_equals: No last remembered size - clientWidth expected 40 but got 100
PASS Early ResizeObserver
PASS Late ResizeObserver
